Bugfix(update,delete):forbiden

+++ webapp/blog/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render,redirect,get_object_or_404
 from django.http import HttpResponse
+from django.contrib.auth.models import User
 from django import forms
 from .models import Post
 from django.views.generic import ListView,DetailView
@@ -9,11 +10,19 @@ from django.shortcuts import render
 
 
 
+
+
 def home(request):
     context = {
-        'posts':Post.objects.all()
+        'posts': Post.objects.all().order_by('-date_posted'),
+        'users': User.objects.all(),
+
     }
+
     return render(request, 'blog/home.html', context)
+
+
+
 
 
 def about(request):
@@ -35,11 +44,16 @@ def post_create(request):
     return render(request, "blog/post_create.html")
 
 
+#  BUG(post_update):fix logic for only author can edit the post
+
 @login_required
 def post_update(request,pk):
     post=get_object_or_404(Post,pk=pk)
+    
+    if request.user != post.author:
+        return HttpResponse("<h1> 403 Forbidden </h1>")
 
-    if request.method=="POST":
+    if request.method=="POST" and request.user==post.author:
         post.title=request.POST.get("title")
         post.content=request.POST.get("content")
         post.save()
@@ -53,16 +67,13 @@ def post_update(request,pk):
 def post_delete(request,pk):
     post=get_object_or_404(Post,pk=pk)
 
-    if request.method=="GET":
+    if request.user != post.author:
+        return HttpResponse("<h1> 403 Forbidden </h1>")
+
+    if request.method=="GET" and request.user==post.author:
        post.delete()
        return redirect("blog-home")
     return redirect("blog-home")
-
-class PostListView(ListView):
-    model=Post
-    context_object_name='posts'
-    template_name='blog/home.html'
-    ordering=['-date_posted']
 
 
 class PostDetailView(DetailView):
